cute_tiled.h: Add cute_tiled_error_cline beside cute_tiled_error_reason
When trying to diagnose a "Unknown identifier found." warning, I
couldn't tell what to look for because this warning appears a bunch of
times in the file so I added cute_tiled_error_cline to print in my logger.

I called it cline to make it clear that it's the line in the C code and
not the line in the json file.