Float properties without decimals are parsed as float (#189)
* Fix MSVC Error C4996

Error	C4996	'strnicmp': The POSIX name for this item is deprecated. Instead,
use the ISO C and C++ conformant name: _strnicmp.

* Fix MSVC Error C4996

Error	C4996	'strnicmp': The POSIX name for this item is deprecated. Instead,
use the ISO C and C++ conformant name: _strnicmp.

* cute_tiled.h: Add cute_tiled_error_cline beside cute_tiled_error_reason

When trying to diagnose a "Unknown identifier found." warning, I
couldn't tell what to look for because this warning appears a bunch of
times in the file so I added cute_tiled_error_cline to print in my logger.

I called it cline to make it clear that it's the line in the C code and
not the line in the json file.

* Ignore vim temp files

* cute_tiled.h: Update to tiled 1.3.3 JSON format > "compressionlevel"

When trying to load a map exported from tiled 1.3.3, I got error
"Unknown identifier found." (cute_tiled.h:2216). The identifier
was "compressionlevel". I just added a case to handle the
identifier and the map loads fine now.

compressionlevel -1 means no compression, just check the value is
indeed -1.

* cute_tiled.h: Put back '('')' in cute_tiled_expect

These were removed in my commit f1394e002, I don't know why they were
removed, they shouldn't have been removed.

* cute_tiled.h: Update to tiled 1.3.3 JSON format > "editorsettings"

Tiled 1.3.3 also adds "editorsettings" to the json file. This is used by
the editor, just skip it.

* Add spaces to respect code style

* Remove space added that shouldn't have been added

* cute_tiled.h: Add myself to contributors, update to v1.04

* cute_tiled.h: Add check for EOF

* cute_tiled.h: Parse float properties without decimals as float