Improve logging info and fix int parsing when it's a float (#203)
* Improve logging info when there's an error

cute_tiled_warning prints the file and line of the error in the json file,
cute_tiled_expect prints the character it found and the character it
expected.
Helps debugging. A warning error now looks like:
WARNING (cute_tiled:__LINE__): Found unexpected ''', expect ':' (my_file.json:75)

* Fix cute_tiled_read_int when the value is a float

Tiled can save values that used to be int, as float. when cute_tiled
expects an int, but there's a float, the map loading fails. This fixes
it.