cute_aseprite_load_from_X option to not deref the color palette when compiling frame->pixels
This is kind of the dumbest solution to a feature that I would like in cute_aseprite. When it is compiling all of the layers together, it would be nice if there was some sort of option to not deref the color palette. I am using the image in indexed mode and it seems like a waste of cpu time to recompile the frames if it just got done looking up all of the palette colors.

I don't think this is the best solution, it changes the api which isn't good. But I can't think of a better way to do it.